requirements.txt
# Парсинг
playwright==1.40.0
aiohttp==3.9.5

# База данных
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Дедупликация
rapidfuzz==3.6.1

# Утилиты
python-dotenv==1.0.0
